// src/types.ts
export type KinokoruItem = {
  id: string;
  timestamp: string;
  name: string;
  job: string;
  background: string;
  category: string;
  level: number;
  title: string;
  description: string;
  benefit: string;
  duration: string;
  price_text: string;
  availability: string;
  achievements: string;
  contact_text: string;
  notes: string;
  slides_url: string;
};